updated the search method
I am trying to move away from the index based approach that I currently use for the list class. This started with changing the search method. Rather than returning the index of the requested item, it returns a pointer to the node of the requested item. If the item is not found, it returns nullptr. This should allow me to use methods like GetNext and others that don't modify the node without having to first search for the node index and then try accessing that node. This approach should be easier to work with and more efficient.