Added forward declaration + made list a friend of node.
I made the linked list class a friend of the node class.

I also got rid of the setters

This allowed me to get rid of the annoying "Node<T> const *" return types for the search methods (since the user, even if they have the address of a non-const node pointer, can't fuck up the list by messing with the node's next value).

I also renamed Entry to CurrentNode (this makes more sense imo) and Prev to PrevNode and Next to NextNode. See Linked_List.hxx for details.

The code still runs with these changes.